code_script_notebooks/python_scripts/company_mgmt/requirements.txt
streamlit==1.23.1
streamlit-agraph==0.0.45
streamlit-pandas-profiling
